requirements.txt
backports.ssl-match-hostname==3.4.0.2
certifi==2015.4.28
cx-Oracle==5.2
Django==1.8.2
django-crispy-forms==1.4.0
django-debug-toolbar==1.3.2
django-filter==0.10.0
django-registration-redux==1.2
djangorestframework==3.1.3
functools32==3.2.3.post1
ipython==3.2.0
Jinja2==2.7.3
jsonschema==2.5.1
Markdown==2.6.2
MarkupSafe==0.23
mistune==0.6
ptyprocess==0.5
Pygments==2.0.2
pyzmq==14.7.0
sqlparse==0.1.15
terminado==0.5
tornado==4.2
wheel==0.24.0
